1. Roperator to calculate Gv when Jacobian=false. 2. Both Z and (phiZ) are stored for Gv when Jacobian = true. 3. phiZ is stored after feedforward. 4. stored idx_phiZ linear indices for bsize instances. 5. If gpu_use=false and without the option Jacobian, Jacobian=true.

+++ MATLAB/cnn/accum.cu
@@ -29,7 +29,7 @@ void __global__ mex_accum(const double *idx, const Tval *val, unsigned long long
 
 	while (i < N)
 	{
-		unsigned long long int vTP_idx = idx[i] - 1;
+		unsigned long long int vTP_idx = (int) idx[i] - 1;
 		if(val[i] != 0.0)
 			atomicAdd(&vTP[vTP_idx], val[i]);
 		i += blockDim.x*gridDim.x;
